Major functionality finished

+++ src/main/java/com/example/javawebapp/model/Bar.java
@@ -3,6 +3,8 @@ package com.example.javawebapp.model;
 import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 
@@ -14,15 +16,15 @@ public class Bar {
     private String id;
     private LocalDate date;
     private String symbol;
-    private Double open;
-    private Double high;
-    private Double low;
-    private Double close;
-    private Double volume;
-    private Double adjClose;
+    private BigDecimal open;
+    private BigDecimal high;
+    private BigDecimal low;
+    private BigDecimal close;
+    private Long volume;
+    private BigDecimal adjClose;
 
-    public Bar(LocalDate date, String symbol, Double open, Double high, Double low, Double close, Double volume,
-               Double adjClose) {
+    public Bar(LocalDate date, String symbol, BigDecimal open, BigDecimal high, BigDecimal low, BigDecimal close,
+               Long volume, BigDecimal adjClose) {
         this.date = date;
         this.symbol = symbol;
         this.open = open;
